Optimize ecp module to remove cache and loader
