--- linkis-engineconn-plugins/linkis-engineconn-plugin-framework/linkis-engineconn-plugin-cache/src/main/java/org/apache/linkis/manager/engineplugin/cache/refresh/RefreshableEngineConnPluginCache.java
@@ -1,59 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.linkis.manager.engineplugin.cache.refresh;
-
-import org.apache.linkis.manager.engineplugin.cache.EngineConnPluginCache;
-import org.apache.linkis.manager.engineplugin.common.loader.entity.EngineConnPluginInfo;
-import org.apache.linkis.manager.engineplugin.common.loader.entity.EngineConnPluginInstance;
-
-public interface RefreshableEngineConnPluginCache extends EngineConnPluginCache {
-
-    /**
-     * Add the refreshListener
-     *
-     * @param refreshListener listener
-     */
-    void addRefreshListener(RefreshListener refreshListener);
-
-    /**
-     * Set refresher
-     *
-     * @param refresher
-     */
-    void setRefresher(PluginCacheRefresher refresher);
-
-    /**
-     * Refresh method
-     *
-     * @param pluginInfo
-     * @param pluginInstance
-     * @throws Exception
-     */
-    void refresh(EngineConnPluginInfo pluginInfo, EngineConnPluginInstance pluginInstance)
-            throws Exception;
-
-    interface RefreshListener {
-
-        /**
-         * On refresh
-         *
-         * @param enginePluginInfo plugin info
-         */
-        void onRefresh(EngineConnPluginInfo enginePluginInfo);
-    }
-}
