+++ linkis-engineconn-plugins/linkis-engineconn-plugin-framework/linkis-engineconn-plugin-server/src/main/java/org/apache/linkis/engineplugin/cache/refresh/PluginCacheRefresher.java
@@ -5,25 +5,23 @@
  * The ASF licenses this file to You under the Apache License, Version 2.0
  * (the "License"); you may not use this file except in compliance with
  * the License.  You may obtain a copy of the License at
- * 
+ *
  *   http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
- 
-package org.apache.linkis.engineplugin.server.loader
 
-import org.apache.linkis.engineplugin.loader.loaders.{DefaultEngineConnPluginLoader, EngineConnPluginsLoader}
+package org.apache.linkis.engineplugin.cache.refresh;
 
+import java.util.concurrent.TimeUnit;
 
-object EngineConnPluginsLoader {
+public interface PluginCacheRefresher {
 
-  private val engineConnPluginsLoader: EngineConnPluginsLoader = new DefaultEngineConnPluginLoader()
+    long interval();
 
-  def getEngineConnPluginsLoader(): EngineConnPluginsLoader = engineConnPluginsLoader
-
+    TimeUnit timeUnit();
 }
